Restructure context package
