+++ data_safe_haven/commands/context/command_group.py
@@ -1,0 +1,5 @@
+"""Command group and entrypoints for managing a DSH context"""
+
+import typer
+
+context_command_group = typer.Typer()
